Rename 'data.py' -> 'dataset.py'
Add usage of partition function to README
