+++ tests/test_dataset.py
@@ -10,7 +10,7 @@ import torch
 import torchvision.transforms as transforms
 from torchvision.datasets import MNIST
 
-from src.data import add_ids, partition_dataset
+from src.dataset import add_ids, partition_dataset
 
 
 class TestVerticalDataset:
